+++ autodocodec-api-usage/test/Autodocodec/YamlSpec.hs
@@ -73,6 +73,8 @@ spec = do
   yamlCodecSpec @VeryComment
   yamlCodecSpec @LegacyValue
   yamlCodecSpec @LegacyObject
+  yamlCodecSpec @Ainur
+  yamlCodecSpec @War
 
 yamlCodecSpec :: forall a. (Show a, Eq a, Typeable a, GenValid a, ToJSON a, FromJSON a, HasCodec a) => Spec
 yamlCodecSpec = describe (nameOf @a) $ do