+++ autodocodec-openapi3/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.1.0] - 2021-12-23
+
+### Added
+
+* `disjointEitherCodec` now no longer generates `additionalProperties = true` and uses `oneOf` instead of `anyOf`.
+
 ## [0.0.0.0] - 2021-11-19
 
 First release.