+++ autodocodec-schema/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.1.0.0] - 2021-12-21
+
+### Added
+
+* Support for `autodocodec-0.0.1.0`.
+
+### Changed
+
+* `ChoiceSchema` has been split into `OneOfSchema` and `AnyOfSchema`
+
 ## [0.0.0.0] - 2021-11-19
 
 First release.