+++ autodocodec-swagger2/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.1.0] - 2021-11-19
+
+### Added
+
+* `disjointEitherCodec` now does not generate `additionalProperties = true`.
+
 ## [0.0.0.0] - 2021-11-19
 
 First release.