+++ autodocodec-yaml/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.0.0] - 2021-12-23
+
+### Changed
+
+* Support for `autodocodec-schema >=0.1.0.0` with comments for `anyOf` and `oneOf`.
+
 ## [0.0.0.0] - 2021-11-19
 
 First release.