+++ autodocodec/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.0.1.0] - 2021-12-23
+
+### Changed
+
+* `EitherCodec` now takes a `Union` to specify whether the union is disjoint or not.
+
+### Added
+
+* `disjointEitherCodec` and `possiblyJointEitherCodec`.
+
 ## [0.0.0.0] - 2021-11-19
 
 First release.